requirements.txt
asgiref==3.7.2
asttokens==2.4.1
boto3==1.34.25
botocore==1.34.25
cffi==1.16.0
cryptography==41.0.7
csvreader==0.0.4
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-crispy-forms==2.1
django-filter==23.5
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
filelock==3.13.1
ipython==8.20.0
jedi==0.19.1
jmespath==1.0.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.24
platformdirs==4.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.0.7
uuid7==0.1.0
virtualenv==20.24.7
wcwidth==0.2.13
